+++ src/Builder.h
@@ -23,7 +23,7 @@ private:
     int32_t m_minerals;
     int32_t m_vespene;
 
-    int32_t m_availableFood;
+    float m_availableFood;
 };
 
 #endif
